Move IBC exports to exports.rs
