+++ packages/std/src/lib.rs
@@ -76,16 +76,13 @@ mod memory; // Used by exports and imports only. This assumes pointers are 32 bi
 
 #[cfg(target_arch = "wasm32")]
 pub use crate::exports::{do_execute, do_instantiate, do_migrate, do_query, do_reply, do_sudo};
-#[cfg(target_arch = "wasm32")]
-pub use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
-
 #[cfg(all(feature = "stargate", target_arch = "wasm32"))]
-mod ibc_exports;
-#[cfg(all(feature = "stargate", target_arch = "wasm32"))]
-pub use crate::ibc_exports::{
+pub use crate::exports::{
     do_ibc_channel_close, do_ibc_channel_connect, do_ibc_channel_open, do_ibc_packet_ack,
     do_ibc_packet_receive, do_ibc_packet_timeout,
 };
+#[cfg(target_arch = "wasm32")]
+pub use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 
 // Exposed for testing only
 // Both unit tests and integration tests are compiled to native code, so everything in here does not need to compile to Wasm.
